[#ENYO-829] serve the ARES IDE using a local Node.js server
This commit also contains the base infrastructure to support many projects in the workspace.  Making the setup of these many projects configurable from the IDE is the topic of anther Areas-related task.

The local IDE server spawn a sub-Node.js interfacing with each project's target storage.  The current storage service access syntax (verbs at the root of the server) prevent feeding these file servers into the main IDE Node.js servers.  This is a limitation that may be addressed in the future.